LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/cross-origin-objects/cross-origin-objects-on-new-window-expected.txt

CONSOLE MESSAGE: line 13: ReferenceError: Can't find variable: get_port

Harness Error (TIMEOUT), message = null
